+++ config/routes.rb
@@ -250,6 +250,9 @@ Rails.application.routes.draw do
     end
   end
 
+  # Project groups
+  resources :groups, only: %i[index show], controller: :project_groups, as: :project_groups
+
   # Service support projects
   constraints(id: VALID_UUID_REGEX) do
     resources :projects, only: %i[index] do