+++ spec/models/project_group_spec.rb
@@ -9,4 +9,35 @@ RSpec.describe ProjectGroup, type: :model do
   describe "associations" do
     it { is_expected.to have_many(:projects) }
   end
+
+  describe "#trust" do
+    it "fetches the trust for the group from the api" do
+      subject = build(:project_group)
+      trust = build(:academies_api_trust)
+
+      mock_successful_api_trust_response(ukprn: subject.trust_ukprn, trust: trust)
+
+      expect(subject.trust).to be trust
+    end
+
+    it "returns an empty trust object with an error when the trust cannot be fetched" do
+      subject = build(:project_group)
+
+      mock_trust_not_found(ukprn: subject.trust_ukprn)
+
+      expect(subject.trust.original_name).to eql "Could not find trust for UKPRN 1234567"
+    end
+
+    it "only calls the API once" do
+      subject = build(:project_group)
+      trust = build(:academies_api_trust)
+
+      api_client = mock_successful_api_trust_response(ukprn: subject.trust_ukprn, trust: trust)
+
+      subject.trust
+      subject.trust
+
+      expect(api_client).to have_received(:get_trust).exactly(1).time
+    end
+  end
 end